handle refresh token

+++ src/hooks/ApiCall.js
@@ -37,30 +37,31 @@ export const ApiCall = async (endpoint, method, token, refreshToken, setToken, s
         if (error.response && error.response.status === 401) {
             // Token has expired, attempt to refresh it
             try {
+                let refetchedData;
                 const refreshLocal = localStorage.getItem("refresh")
-                console.log(refreshLocal)
                 let response = await refreshAccessToken(refreshLocal);
-                
-                response.access? '': response = await refreshAccessToken(refreshLocal);
-
+                if (response === null){
+                    return ToastMessage("Ooops", "An error occured")
+                }
                 const {access, refresh} = response
                 if (access) {
                     setToken(access);
                     setRefresh(refresh);
                     localStorage.setItem("site", access);
                     localStorage.setItem("refresh", refresh);
+
                     // Retry the original request with the new token
                     headers["Authorization"] = `Bearer ${access}`;
                     if (method === 'post') {
-                        res = await api.post(endpoint, data, { headers });
+                        refetchedData = await api.post(endpoint, data, { headers });
                     } else if (method === 'get') {
-                        res = await api.get(endpoint, { headers });
+                        refetchedData = await api.get(endpoint, { headers });
                     }else if (method === 'delete') {
-                        res = await api.delete(endpoint, { headers });
+                        refetchedData = await api.delete(endpoint, { headers });
                     }else if (method === 'delete'){
-                        res = await api.put(endpoint, data, { headers });
+                        refetchedData = await api.put(endpoint, data, { headers });
                     }
-                    return res;
+                    return refetchedData
                 } else {
                     ToastMessage("error", "You are not authorized to perform this function");
                 }
